Run tests in Docker during image build
We prefer to run tests in Docker during image build for several reasons:

1. With such approach there is no way the bad image (with tests failing) could be created.

2. We avoid multiple builds for various needs. If tests are launched outside Docker, one more build outside container is required.
Another option is to avoid build inside container at all and just copy build artefacts to the image.

3. We keep the whole build pipeline (restore/build/test/publish) in one place. The switch to another CI solution is trivial: single docker build command.

As part of this transition, we remove automatic launch of docker-compose from integration needs.
Now all dependencies must be up and running when docker build is invoked (so that IT launched inside container have all required dependencies).
There is no way to spin-up those dependencies during the docker build command, so it must be outer invocation.